Moved login presenter due to git error
